add fine-tuning for key point generation
this adds a module that adds a header onto a pre-trained model that, for each point in a patch, attempts to generate a set of end points for a line segment. the idea is that the model will hopefully learn the start and end points for each particle instance such that we can cluster these start/end points to define instances of 1D lines in the pointcloud. this is similar to what VoteNet does.