add PPN code & training module
PPN teaches a MLP header to, for each group token, propose a set of endpoints as well as the number of associated particles. these endpoints are then aggregated via voxel clustering, and then matched to true endpoints via Hungarian matching. then loss between the matches (L1 for coords, CE for classification)

note that the proposals are parametrized as offsets from the center of each group, in units of the group radius.