Fix sphere nms! + fast random start point for fps & readability improvements
sphere nms output was incorrectly used to reduce the number of point centers to use. it was (stupidly) implicitly assumed that retain returned a homogeneous array of [1,1,1,1,...,0,0,0] and so the first N elements of the groups were taken, where N is the number of 1's in the retain array. Obviously `retain` looks more like [0,0,1,0,...,1,0,1,0], i.e. random and so this is very wrong.